+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased](https://github.com/dalance/svlint/compare/v0.3.0...Unreleased) - ReleaseDate
 
+* [Added] --ignore-include option #[10](https://github.com/dalance/svlint/issues/10)
 * [Changed] --define can have value like `--define NAME=VALUE` [#9](https://github.com/dalance/svlint/issues/9)
 
 ## [v0.3.0](https://github.com/dalance/svlint/compare/v0.2.21...v0.3.0) - 2020-01-23